LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.arc.nonfinite.worker-expected.txt


PASS arc() with Infinity/NaN is ignored 
